// src/trainer-profile/dtos/trainer-profile-detailed-options.dto.ts
export class TrainerProfileDetailedOptions {
  includeLevel?: boolean;
  includeUser?: boolean;
  includeAgeGroup?: boolean;
  includeRegion?: boolean;
  includeGallery?: boolean;
  includeCertificates?: boolean;
  includePackages?: boolean;
  includeFields?: boolean;
  includeSports?: boolean;
  includeFeedbacks?: boolean;
  includeRatingNumber?: boolean;
}